Begin reorganisation of shared/v4/v6 ICMP code
