+++ lib/opte/src/engine/headers.rs
@@ -11,9 +11,9 @@ use super::geneve::GeneveHdr;
 use super::geneve::GeneveMeta;
 use super::geneve::GeneveMod;
 use super::geneve::GenevePush;
-use super::icmpv6::IcmpHdr;
-use super::icmpv6::Icmpv4Meta;
-use super::icmpv6::Icmpv6Meta;
+use super::icmp::IcmpHdr;
+use super::icmp::Icmpv4Meta;
+use super::icmp::Icmpv6Meta;
 use super::ip4::Ipv4Hdr;
 use super::ip4::Ipv4Meta;
 use super::ip4::Ipv4Mod;
