+++ lib/opte/src/engine/icmp/v4.rs
@@ -4,25 +4,29 @@
 
 // Copyright 2023 Oxide Computer Company
 
-//! ICMP headers.
-use super::ether::EtherHdr;
-use super::ether::EtherMeta;
-use super::ether::EtherType;
-use super::ip4::Ipv4Hdr;
-use super::ip4::Ipv4Meta;
-use super::packet::Packet;
-use super::packet::PacketMeta;
-use super::packet::PacketRead;
-use super::packet::PacketReader;
-use super::predicate::DataPredicate;
-use super::predicate::EtherAddrMatch;
-use super::predicate::IpProtoMatch;
-use super::predicate::Ipv4AddrMatch;
-use super::predicate::Predicate;
-use super::rule::AllowOrDeny;
-use super::rule::GenErr;
-use super::rule::GenPacketResult;
-use super::rule::HairpinAction;
+//! ICMPv4 headers and processing.
+
+use super::IcmpMeta;
+use crate::engine::ether::EtherHdr;
+use crate::engine::ether::EtherMeta;
+use crate::engine::ether::EtherType;
+use crate::engine::icmp::HeaderActionModify;
+use crate::engine::icmp::UlpMetaModify;
+use crate::engine::ip4::Ipv4Hdr;
+use crate::engine::ip4::Ipv4Meta;
+use crate::engine::packet::Packet;
+use crate::engine::packet::PacketMeta;
+use crate::engine::packet::PacketRead;
+use crate::engine::packet::PacketReader;
+use crate::engine::predicate::DataPredicate;
+use crate::engine::predicate::EtherAddrMatch;
+use crate::engine::predicate::IpProtoMatch;
+use crate::engine::predicate::Ipv4AddrMatch;
+use crate::engine::predicate::Predicate;
+use crate::engine::rule::AllowOrDeny;
+use crate::engine::rule::GenErr;
+use crate::engine::rule::GenPacketResult;
+use crate::engine::rule::HairpinAction;
 use alloc::vec::Vec;
 use core::fmt;
 use core::fmt::Display;
@@ -33,8 +37,42 @@ use serde::Serialize;
 use smoltcp::phy::Checksum;
 use smoltcp::phy::ChecksumCapabilities as Csum;
 use smoltcp::wire;
+use smoltcp::wire::Icmpv4Message;
 use smoltcp::wire::Icmpv4Packet;
 use smoltcp::wire::Icmpv4Repr;
+
+pub type Icmpv4Meta = IcmpMeta<MessageType>;
+
+impl Icmpv4Meta {
+    /// Extract an ID from the body of an ICMPv4 packet to use as a
+    /// pseudo port for flow differentiation.
+    ///
+    /// This method returns `None` for any non-echo packets.
+    #[inline]
+    pub fn echo_id(&self) -> Option<u16> {
+        match self.msg_type.inner {
+            Icmpv4Message::EchoRequest | Icmpv4Message::EchoReply => {
+                Some(u16::from_be_bytes(self.body_echo().id))
+            }
+            _ => None,
+        }
+    }
+}
+
+impl HeaderActionModify<UlpMetaModify> for Icmpv4Meta {
+    fn run_modify(&mut self, spec: &UlpMetaModify) {
+        let Some(new_id) = spec.icmp_id else {
+            return;
+        };
+
+        if self.echo_id().is_none() {
+            return;
+        }
+
+        let mut echo_data = self.body_echo_mut();
+        echo_data.id = new_id.to_be_bytes();
+    }
+}
 
 impl HairpinAction for IcmpEchoReply {
     fn implicit_preds(&self) -> (Vec<Predicate>, Vec<DataPredicate>) {
@@ -190,5 +228,40 @@ impl From<u8> for MessageType {
 impl Display for MessageType {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.inner)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use crate::engine::checksum::Checksum as OpteCsum;
+    use crate::engine::headers::RawHeader;
+    use crate::engine::icmp::IcmpHdr;
+    use crate::engine::icmp::IcmpHdrRaw;
+    use smoltcp::wire::Icmpv4Packet;
+    use smoltcp::wire::Icmpv4Repr;
+
+    use super::*;
+
+    #[test]
+    fn icmp4_body_csum_equals_body() {
+        let data = b"reunion\0";
+        let mut body_csum = OpteCsum::default();
+        body_csum.add_bytes(data);
+
+        let mut cksum_cfg = Csum::ignored();
+        cksum_cfg.icmpv4 = Checksum::Both;
+
+        let test_pkt = Icmpv4Repr::EchoRequest { ident: 7, seq_no: 7777, data };
+        let mut out = vec![0u8; test_pkt.buffer_len()];
+        let mut packet = Icmpv4Packet::new_unchecked(&mut out);
+        test_pkt.emit(&mut packet, &cksum_cfg);
+
+        let src = &mut out[..IcmpHdr::SIZE];
+        let icmp = IcmpHdr { base: IcmpHdrRaw::new_mut(src).unwrap() };
+
+        assert_eq!(
+            Some(body_csum.finalize()),
+            icmp.csum_minus_hdr().map(|mut v| v.finalize())
+        );
     }
 }
