+++ lib/opte/src/engine/predicate.rs
@@ -10,8 +10,8 @@ use super::dhcp::MessageType as DhcpMessageType;
 use super::dhcpv6::MessageType as Dhcpv6MessageType;
 use super::ether::EtherType;
 use super::headers::IpMeta;
-use super::icmp::MessageType as IcmpMessageType;
-use super::icmpv6::MessageType as Icmpv6MessageType;
+use super::icmp::v4::MessageType as IcmpMessageType;
+use super::icmp::v6::MessageType as Icmpv6MessageType;
 use super::ip4::Ipv4Addr;
 use super::ip4::Ipv4Cidr;
 use super::ip4::Ipv4Meta;
