+++ lib/oxide-vpc/src/engine/gateway/icmpv6.rs
@@ -13,9 +13,9 @@ use core::result::Result;
 use opte::api::Direction;
 use opte::api::Ipv6Addr;
 use opte::api::OpteError;
-use opte::engine::icmpv6::Icmpv6EchoReply;
-use opte::engine::icmpv6::NeighborAdvertisement;
-use opte::engine::icmpv6::RouterAdvertisement;
+use opte::engine::icmp::v6::Icmpv6EchoReply;
+use opte::engine::icmp::v6::NeighborAdvertisement;
+use opte::engine::icmp::v6::RouterAdvertisement;
 use opte::engine::layer::Layer;
 use opte::engine::predicate::DataPredicate;
 use opte::engine::rule::Action;
